requirements.txt
# Tembo AI - Backend Dependencies
# Lightweight requirements for Render free tier

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
gunicorn>=21.0.0

# LLM - Groq (free)
groq>=0.4.0

# PostgreSQL + pgvector
psycopg2-binary>=2.9.0

# Environment & Data
python-dotenv>=1.0.0
pydantic>=2.0.0

# Local embeddings (small model, ~90MB)
sentence-transformers>=2.2.0

# For embedding processing
numpy>=1.24.0
